requirements.txt
Django==1.8.4
MySQL-python==1.2.5
Pillow==2.9.0
argparse==1.2.1
decorator==4.0.2
djangorestframework==3.2.3
ipdb==0.8.1
ipython==4.0.0
ipython-genutils==0.1.0
path.py==8.1
pexpect==3.3
pickleshare==0.5
simplegeneric==0.8.1
traitlets==4.0.0
wsgiref==0.1.2
